there is a repeated error here.
there are two sets of dots:
1) static
2) dynamic

you always get rid of the wrong one.

now we're back to having the static one only.

analyze the code, and propose alternative ways to proceed, I propose no dots at all, but lines that connect to each other where the values would be